# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse:
    properties:
      code:
        description: |-
          Code is a machine-readable status identifier for the client to act on.
          Use SCREAMING_SNAKE_CASE.
          Example: "CREATED" | "NOT_FOUND" | "VALIDATION_ERROR" | "INTERNAL_ERROR"
        type: string
      data:
        description: |-
          Data carries the operation's result payload.
          Omitted from the JSON output when nil.
      message:
        description: |-
          Message is a human-readable description of the outcome.
          On 4xx/5xx responses this field is masked with a reference ID before reaching the client.
          Example: "candidate created successfully"
        type: string
      success:
        description: |-
          Success indicates whether the operation completed without error.
          Example: true
        type: boolean
    type: object
...
  internal_question.CreateRequest:
    properties:
      job_posting_id:
        type: string
      order_index:
        type: integer
      text:
        type: string
    required:
    - job_posting_id
    - text
    type: object
  internal_question.Question:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      created_by_name:
        type: string
      created_ip_address:
        type: string
      id:
        type: string
      job_posting_id:
        type: string
      order_index:
        type: integer
      text:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      updated_by_name:
        type: string
      updated_ip_address:
        type: string
    type: object
  internal_question.UpdateRequest:
    properties:
      order_index:
        type: integer
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Nexus Team
  description: |-
    REST API for the Nexus AI-driven interview platform.

    Authentication: protected endpoints accept a JWT either as an `Authorization: Bearer {token}` header
    or as the HttpOnly `access_token` cookie set by the auth endpoints. In Swagger UI, click Authorize and enter `Bearer {token}`.

    Every response is wrapped in the standard envelope `{ success, message, code, data }`.
    On 4xx/5xx responses the `message` field may be masked with a reference ID.
  title: Nexus Interview API
  version: "1.0"
paths:
...
  /questions:
    post:
      consumes:
      - application/json
      description: |-
        Creates a question bound to the given `job_posting_id`. `order_index` controls the position of the
        question inside the interview (lower first, default 0). The generated UUID is returned in `data.id`.
      parameters:
      - description: Question to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_question.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: CREATED
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_question.Question'
              type: object
        "400":
          description: VALIDATION_ERROR — missing job_posting_id or text
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "401":
          description: UNAUTHORIZED
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
      security:
      - BearerAuth: []
      summary: Create question
      tags:
      - Questions
  /questions/{id}:
    delete:
      description: |-
        Marks the question as deleted (soft delete). It no longer appears in job posting listings.
        Deleting an already-deleted or unknown id yields 404.
      parameters:
      - description: Question UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DELETED
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "400":
          description: VALIDATION_ERROR — empty id
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "401":
          description: UNAUTHORIZED
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "404":
          description: NOT_FOUND
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
      security:
      - BearerAuth: []
      summary: Delete question
      tags:
      - Questions
    get:
      description: Fetches one question by its UUID. Soft-deleted questions are treated
        as not found.
      parameters:
      - description: Question UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_question.Question'
              type: object
        "400":
          description: VALIDATION_ERROR — empty id
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "401":
          description: UNAUTHORIZED
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "404":
          description: NOT_FOUND
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
      security:
      - BearerAuth: []
      summary: Get question
      tags:
      - Questions
    put:
      consumes:
      - application/json
      description: |-
        Updates `text` and `order_index` of an existing question. The owning job posting cannot be changed;
        delete and recreate the question to move it. Returns the full updated record.
      parameters:
      - description: Question UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_question.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_question.Question'
              type: object
        "400":
          description: VALIDATION_ERROR — empty id or malformed body
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "401":
          description: UNAUTHORIZED
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "404":
          description: NOT_FOUND
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
      security:
      - BearerAuth: []
      summary: Update question
      tags:
      - Questions
  /questions/job-posting/{jobPostingId}:
    get:
      description: |-
        Returns every non-deleted question attached to the given job posting, ordered by `order_index`.
        An unknown job posting yields an empty list, not 404.
      parameters:
      - description: Job posting UUID
        format: uuid
        in: path
        name: jobPostingId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_question.Question'
                  type: array
              type: object
        "400":
          description: VALIDATION_ERROR — empty job posting id
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "401":
          description: UNAUTHORIZED
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/github_com_fathdemr_nexus-interview_pkg_httputil.BaseServiceResponse'
      security:
      - BearerAuth: []
      summary: List questions of a job posting
      tags:
      - Questions
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token as: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...